+++ v3/cmd/wails3/main.go
@@ -57,10 +57,8 @@ func main() {
 	generate.NewSubCommandFunction(".desktop", "Generate .desktop file", commands.GenerateDotDesktop)
 	generate.NewSubCommandFunction("appimage", "Generate Linux AppImage", commands.GenerateAppImage)
 
-	plugin := app.NewSubCommand("plugin", "Plugin tools")
-	//plugin.NewSubCommandFunction("list", "List plugins", commands.PluginList)
-	plugin.NewSubCommandFunction("init", "Initialise a new plugin", commands.PluginInit)
-	//plugin.NewSubCommandFunction("add", "Add a plugin", commands.PluginAdd)
+	plugin := app.NewSubCommand("service", "Service tools")
+	plugin.NewSubCommandFunction("init", "Initialise a new service", commands.ServiceInit)
 	tool := app.NewSubCommand("tool", "Various tools")
 	tool.NewSubCommandFunction("checkport", "Checks if a port is open. Useful for testing if vite is running.", commands.ToolCheckPort)
 	tool.NewSubCommandFunction("watcher", "Watches files and runs a command when they change", commands.Watcher)