+++ v3/internal/assetserver/assetserver.go
@@ -1,14 +1,11 @@
 package assetserver
 
 import (
-	"embed"
 	"fmt"
-	"io/fs"
 	"net"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"path"
 	"strings"
 	"time"
 )
@@ -16,7 +13,7 @@ import (
 const (
 	webViewRequestHeaderWindowId   = "x-wails-window-id"
 	webViewRequestHeaderWindowName = "x-wails-window-name"
-	pluginPrefix                   = "/wails/plugin"
+	servicePrefix                  = "wails/services"
 )
 
 type RuntimeHandler interface {
@@ -28,16 +25,14 @@ type AssetServer struct {
 
 	handler http.Handler
 
-	//pluginScripts map[string]string
+	services map[string]http.Handler
 
 	assetServerWebView
-	pluginAssets map[string]fs.FS
 }
 
 func NewAssetServer(options *Options) (*AssetServer, error) {
 	result := &AssetServer{
-		options:      options,
-		pluginAssets: make(map[string]fs.FS),
+		options: options,
 	}
 
 	userHandler := options.Handler
@@ -113,37 +108,31 @@ func (a *AssetServer) serveHTTP(rw http.ResponseWriter, req *http.Request, userH
 
 	default:
 
-		// Check if this is a plugin asset
-		if !strings.HasPrefix(reqPath, pluginPrefix) {
-			userHandler.ServeHTTP(rw, req)
-			return
-		}
-
-		// Ensure there is 4 parts to the reqPath
-		parts := strings.SplitN(reqPath, "/", 5)
-		if len(parts) < 5 {
-			rw.WriteHeader(http.StatusNotFound)
-			return
-		}
-
-		// Get the first 3 parts of the reqPath
-		pluginPath := "/" + path.Join(parts[1], parts[2], parts[3])
-		// Get the remaining part of the reqPath
-		fileName := parts[4]
-
-		// Check if this is a registered plugin asset
-		if assetFS, ok := a.pluginAssets[pluginPath]; ok {
-			// Check if the file exists
-			file, err := fs.ReadFile(assetFS, fileName)
-			if err != nil {
-				a.serveError(rw, err, "Unable to read file %s", reqPath)
+		// Check if the path matches the keys in the services map
+		for route, handler := range a.services {
+			if strings.HasPrefix(reqPath, route) {
+				req.URL.Path = strings.TrimPrefix(reqPath, route)
+				handler.ServeHTTP(rw, req)
 				return
 			}
-			a.writeBlob(rw, reqPath, file)
-		} else {
-			userHandler.ServeHTTP(rw, req)
 		}
+
+		// Check if it can be served by the user-provided handler
+		if !strings.HasPrefix(reqPath, servicePrefix) {
+			userHandler.ServeHTTP(rw, req)
+			return
+		}
+
+		rw.WriteHeader(http.StatusNotFound)
+		return
 	}
+}
+
+func (a *AssetServer) AttachServiceHandler(prefix string, handler http.Handler) {
+	if a.services == nil {
+		a.services = make(map[string]http.Handler)
+	}
+	a.services[prefix] = handler
 }
 
 func (a *AssetServer) writeBlob(rw http.ResponseWriter, filename string, blob []byte) {
@@ -157,30 +146,6 @@ func (a *AssetServer) serveError(rw http.ResponseWriter, err error, msg string, 
 	args = append(args, err)
 	a.options.Logger.Error(msg+":", args...)
 	rw.WriteHeader(http.StatusInternalServerError)
-}
-
-//func (a *AssetServer) AddPluginScript(pluginName string, script string) {
-//	if a.pluginScripts == nil {
-//		a.pluginScripts = make(map[string]string)
-//	}
-//	pluginName = strings.ReplaceAll(pluginName, "/", "_")
-//	pluginName = html.EscapeString(pluginName)
-//	pluginScriptName := fmt.Sprintf("/wails/plugin/%s.js", pluginName)
-//	a.pluginScripts[pluginScriptName] = script
-//}
-
-func (a *AssetServer) AddPluginAssets(pluginPath string, vfs fs.FS) error {
-	pluginPath = path.Join(pluginPrefix, pluginPath)
-	_, exists := a.pluginAssets[pluginPath]
-	if exists {
-		return fmt.Errorf("plugin path already exists: %s", pluginPath)
-	}
-	if embedFs, isEmbedFs := vfs.(embed.FS); isEmbedFs {
-		rootFolder, _ := findEmbedRootPath(embedFs)
-		vfs, _ = fs.Sub(vfs, path.Clean(rootFolder))
-	}
-	a.pluginAssets[pluginPath] = vfs
-	return nil
 }
 
 func GetStartURL(userURL string) (string, error) {