+++ v3/internal/go-common-file-dialog/cfd/DialogConfig.go
@@ -3,6 +3,7 @@
 package cfd
 
 import (
+	"fmt"
 	"os"
 )
 
@@ -114,6 +115,10 @@ func (config *DialogConfig) apply(dialog Dialog) (err error) {
 		}
 
 		if config.SelectedFileFilterIndex != 0 {
+			if config.SelectedFileFilterIndex > uint(len(fileFilters)) {
+				err = fmt.Errorf("selected file filter index out of range")
+				return
+			}
 			err = dialog.SetSelectedFileFilterIndex(config.SelectedFileFilterIndex)
 			if err != nil {
 				return