+++ v3/pkg/application/application_darwin.go
@@ -221,7 +221,7 @@ func (m *macosApp) setApplicationMenu(menu *Menu) {
 
 func (m *macosApp) run() error {
 	// Add a hook to the ApplicationDidFinishLaunching event
-	m.parent.On(events.Mac.ApplicationDidFinishLaunching, func(*Event) {
+	m.parent.OnApplicationEvent(events.Mac.ApplicationDidFinishLaunching, func(*ApplicationEvent) {
 		C.setApplicationShouldTerminateAfterLastWindowClosed(C.bool(m.parent.options.Mac.ApplicationShouldTerminateAfterLastWindowClosed))
 		C.setActivationPolicy(C.int(m.parent.options.Mac.ActivationPolicy))
 		C.activateIgnoringOtherApps()