+++ v3/pkg/application/webview_window_linux.go
@@ -2,8 +2,7 @@
 
 package application
 
-import "C"
-import (
+import ( 
 	"fmt"
 	"time"
 
@@ -312,7 +311,7 @@ func (w *linuxWebviewWindow) run() {
 	}
 
 	// Ignore mouse events if requested
-	w.setIgnoreMouseEvents(options.IgnoreMouseEvents)
+	w.setIgnoreMouseEvents(w.parent.options.IgnoreMouseEvents)
 
 	startURL, err := assetserver.GetStartURL(w.parent.options.URL)
 	if err != nil {
@@ -320,7 +319,7 @@ func (w *linuxWebviewWindow) run() {
 	}
 
 	w.setURL(startURL)
-	w.parent.On(events.Linux.WindowLoadChanged, func(_ *WindowEvent) {
+	w.parent.OnWindowEvent(events.Linux.WindowLoadChanged, func(_ *WindowEvent) {
 		if w.parent.options.JS != "" {
 			w.execJS(w.parent.options.JS)
 		}
@@ -329,19 +328,19 @@ func (w *linuxWebviewWindow) run() {
 			w.execJS(js)
 		}
 	})
-	w.parent.On(events.Linux.WindowFocusIn, func(e *WindowEvent) {
+	w.parent.OnWindowEvent(events.Linux.WindowFocusIn, func(e *WindowEvent) {
 		w.parent.emit(events.Common.WindowFocus)
 	})
-	w.parent.On(events.Linux.WindowFocusOut, func(e *WindowEvent) {
+	w.parent.OnWindowEvent(events.Linux.WindowFocusOut, func(e *WindowEvent) {
 		w.parent.emit(events.Common.WindowLostFocus)
 	})
-	w.parent.On(events.Linux.WindowDeleteEvent, func(e *WindowEvent) {
+	w.parent.OnWindowEvent(events.Linux.WindowDeleteEvent, func(e *WindowEvent) {
 		w.parent.emit(events.Common.WindowClosing)
 	})
-	w.parent.On(events.Linux.WindowDidMove, func(e *WindowEvent) {
+	w.parent.OnWindowEvent(events.Linux.WindowDidMove, func(e *WindowEvent) {
 		w.parent.emit(events.Common.WindowDidMove)
 	})
-	w.parent.On(events.Linux.WindowDidResize, func(e *WindowEvent) {
+	w.parent.OnWindowEvent(events.Linux.WindowDidResize, func(e *WindowEvent) {
 		w.parent.emit(events.Common.WindowDidResize)
 	})
 
@@ -405,11 +404,5 @@ func (w *linuxWebviewWindow) isIgnoreMouseEvents() bool {
 }
 
 func (w *linuxWebviewWindow) setIgnoreMouseEvents(ignore bool) {
-	w.ignoreMouseEvents = ignore
-
-	if ignore {
-		C.gtk_widget_set_events((*C.GtkWidget)(unsafe.Pointer(w.window)), C.GDK_ENTER_NOTIFY_MASK|C.GDK_LEAVE_NOTIFY_MASK)
-	} else {
-		C.gtk_widget_set_events((*C.GtkWidget)(unsafe.Pointer(w.window)), C.GDK_ALL_EVENTS_MASK)
-	}
+	w.ignoreMouse(w.ignoreMouseEvents)
 }